+++ empire/server/modules/bof/clipboard_window_inject.py
@@ -45,10 +45,13 @@ class Module:
 
         params_dict = {
             "Architecture": "x64",
-            "Shellcode": f"-b:{encoded_shellcode}",
-            "ProcessID": f"-i:{pid}",
+            "Shellcode": encoded_shellcode,
+            "ProcessID": pid,
         }
 
         return main_menu.modulesv2.generate_script_bof(
-            module=module, params=params_dict, obfuscate=obfuscate
+            module=module,
+            params=params_dict,
+            obfuscate=obfuscate,
+            skip_params=True,
         )