+++ empire/server/modules/bof/situational_awareness/netshares.py
@@ -13,10 +13,13 @@ class Module:
     ):
         params_dict = {
             "Architecture": params["Architecture"],
-            "Hostname": f"-Z:{params['Hostname']}",
-            "Flag": "-i:0",
+            "Hostname": params["Hostname"],
+            "Flag": "0",
         }
 
         return main_menu.modulesv2.generate_script_bof(
-            module=module, params=params_dict, obfuscate=obfuscate
+            module=module,
+            params=params_dict,
+            obfuscate=obfuscate,
+            skip_params=True,
         )