+++ empire/server/modules/bof/situational_awareness/windowlist.py
@@ -13,9 +13,12 @@ class Module:
     ):
         params_dict = {
             "Architecture": params["Architecture"],
-            "All": f"-i:{'1' if params.get('all') == 'true' else '0'}",
+            "All": {"1" if params.get("all") == "true" else "0"},
         }
 
         return main_menu.modulesv2.generate_script_bof(
-            module=module, params=params_dict, obfuscate=obfuscate
+            module=module,
+            params=params_dict,
+            obfuscate=obfuscate,
+            skip_params=True,
         )