+++ empire/server/utils/file_util.py
@@ -28,7 +28,9 @@ def remove_file(path: str) -> None:
         os.remove(path)
 
 
-def run_as_user(command, user=None, cwd=None):
+def run_as_user(  # noqa: PLR0913
+    command, user=None, cwd=None, capture_output=False, check=True, text=True
+):
     """
     Runs a command as a specified user or the user who invoked sudo.
     If no user is specified and the script is not run with sudo, it runs as the current user.
@@ -36,6 +38,11 @@ def run_as_user(command, user=None, cwd=None):
     Args:
         command (list): The command to run, specified as a list of strings.
         user (str, optional): The username to run the command as. Defaults to None.
+        cwd (str, optional): The working directory for the command. Defaults to None.
+        capture_output (bool, optional): Whether to capture and return the command's output. Defaults to False.
+
+    Returns:
+        str or None: The output of the command if capture_output is True, otherwise None.
     """
     try:
         if user is None:
@@ -43,11 +50,25 @@ def run_as_user(command, user=None, cwd=None):
 
         command_with_user = ["sudo", "-u", user, *command] if user else command
 
-        subprocess.run(command_with_user, check=True, cwd=cwd)
+        result = subprocess.run(
+            command_with_user,
+            check=check,
+            cwd=cwd,
+            text=text,
+            capture_output=capture_output,
+        )
 
         log.debug("Command executed successfully: %s", " ".join(map(str, command)))
 
+        if capture_output:
+            return result.stdout.strip()
+        return None
+
     except subprocess.CalledProcessError as e:
-        # Log the error details
         log.error("Failed to execute command: %s", e, exc_info=True)
         log.error("Try running the command manually: %s", " ".join(command))
+        if e.stdout:
+            log.error("Command output: %s", e.stdout)
+        if e.stderr:
+            log.error("Command error output: %s", e.stderr)
+        raise